Add `validate` to direct bindings (#5048)
This PR creates `validate` in direct bindings. It is the same
[validate](https://github.com/NVIDIA/Fuser/blob/main/python/nvfuser/__init__.py#L580-L625)
in legacy bindings.

* Add `validate_with_auto_inferred_outputs` and `get_val_tolerances` to
`FusionExecutorCache` pybind11 class.
* Move all tests from `tests/python/test_repro.py` to
`tests/python/direct/test_repro.py`.
* The tests no longer use `validate_with_auto_inferred_outputs`
directly.
* Removed `define_vector` from `test_issue4670` because it isn't used in
direct bindings.
