+++ csrc/validator_utils.h
@@ -83,7 +83,7 @@ ExpressionEvaluator bindInputsAndLaunchParams(
     const KernelArgumentHolder& aten_inputs,
     const LaunchParams& launch_constraints);
 
-std::vector<std::pair<double, double>> get_val_constants(
+NVF_API std::vector<std::pair<double, double>> getValTolerances(
     Fusion* fusion,
     const KernelArgumentHolder& aten_inputs,
     const LaunchParams& lparams = LaunchParams(),
@@ -125,8 +125,8 @@ void testValidate(
     const LaunchParams& lparams = LaunchParams(),
     const ValidationConstants& tolerances = ValidationConstants());
 
-// The variant withtout line number and file name.
-void testValidate(
+// The variant without line number and file name.
+NVF_API void testValidate(
     Fusion* fusion,
     const KernelArgumentHolder& fusion_outputs,
     const KernelArgumentHolder& aten_inputs);
