+++ python/python_frontend/fusion_definition.cpp
@@ -733,7 +733,7 @@ std::vector<Tensor> FusionDefinition::tensors() {
 
 std::vector<std::pair<double, double>> FusionDefinition::getValTolerances(
     const KernelArgumentHolder& args) {
-  return get_val_constants(preschedFusion(), args);
+  return nvfuser::getValTolerances(preschedFusion(), args);
 }
 
 void FusionDefinition::validate_with_auto_inferred_outputs(
